pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dsff"
version = "1.0.0"
description = "DS FolderFit — 스마트 폴더 정리 도구"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Arti(AX)", email = "hrlee@shinhan.com" },
]
keywords = ["file-organizer", "folder-cleanup", "duplicate-detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "xxhash>=3.0",
    "filetype>=1.2",
    "send2trash>=1.8",
    "watchdog>=3.0",
    "platformdirs>=4.0",
    "humanize>=4.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-tmp-files>=0.0.2",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
dsff = "dsff.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dsff"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.black]
target-version = ["py39"]
line-length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
